api/requirements.txt
fastapi>=0.95.1
uvicorn[standard]>=0.22.0
pydantic>=1.10.7
typing-extensions>=4.5.0

lingua-language-detector
cutlet
fugashi[unidic-lite]
deepl
httpx